+++ packages/waltz-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # waltz-icons
 
+## 0.0.19
+
+### Patch Changes
+
+- Minor fixes
+
 ## 0.0.18
 
 ### Patch Changes